+++ pyobjc/Modules/objc/objc-runtime-apple.h
@@ -70,8 +70,8 @@ PyObjCRT_ClassAddMethodList(Class cls, struct objc_method_list* lst)
 }
 
 
-extern struct objc_method_list* PyObjCRT_AllocMethodList(int);
-extern struct objc_protocol_list* PyObjCRT_AllocProtocolList(int);
+extern struct objc_method_list* PyObjCRT_AllocMethodList(Py_ssize_t);
+extern struct objc_protocol_list* PyObjCRT_AllocProtocolList(Py_ssize_t);
 
 typedef Method PyObjCRT_Method_t;
 typedef Ivar PyObjCRT_Ivar_t;
