+++ pyobjc/Modules/objc/objc-runtime-gnu.h
@@ -117,8 +117,8 @@ PyObjCRT_InitMethod(Method_t m, SEL name, const char* types, IMP imp)
 }
 
 
-extern MethodList_t PyObjCRT_AllocMethodList(int);
-extern struct objc_protocol_list* PyObjCRT_AllocProtocolList(int);
+extern MethodList_t PyObjCRT_AllocMethodList(Py_ssize_t);
+extern struct objc_protocol_list* PyObjCRT_AllocProtocolList(Py_ssize_t);
 
 
 typedef Method_t PyObjCRT_Method_t;
